add helpers for accessing kb rules
