+++ polar-core/src/rules.rs
@@ -162,7 +162,9 @@ mod tests {
         polar.load_str(r#"f(1, 3, {c: "z"});"#).unwrap();
 
         let kb = polar.kb.read().unwrap();
-        let generic_rule = kb.rules.get(&sym!("f")).unwrap();
+        let generic_rule = kb
+            .lookup_rule(sym!("f").into(), Path::default_path())
+            .unwrap();
         let index = &generic_rule.index;
         assert!(index.rules.is_empty());
